requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
babel==2.17.0
boto3==1.42.33
botocore==1.42.33
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==6.0.1
django-cors-headers==4.9.0
django-money==3.5.4
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
GDAL==3.10.3
h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jmespath==1.1.0
language_tool_python==3.2.2
Markdown==3.10
MarkupSafe==3.0.3
minio==7.2.20
msgpack==1.1.2
packaging==25.0
psutil==7.2.1
psycopg==3.3.2
psycopg-binary==3.3.2
pubcontrol==3.5.0
py-dateutil==2.2
py-moneyed==3.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
s3transfer==0.16.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
toml==0.10.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wheel==0.45.1
yaml-1.3==0.1.0
zope.interface==8.2
